Extended interfaces for upgrades, tests, inventory

+++ contracts/upgrades-registry/test/interfaces/ITest.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.0;
 
 interface ITest {
-    function test() external pure returns (uint);
+    function test() external pure returns (uint256);
 }
